pyproject.toml
[tool.poetry]
name = "ace_pipeline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.1"
python-casacore = "^3.5.1"
tqdm = "^4.64.0"
astropy = "^5.1"
h5py = "^3.7.0"
importlib = "^1.0.4"
importlib-metadata = "^4.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
